Add files via upload
Syndication support added.

max_depth parameter added to control how far into site hierarchy we want to go.
Crawler budget parameter, max_urls added.
Update to linkbank to scale in 256 batches only, instead of exponent. Slightly slower, more reliable.
Added a sys.exit(0) on successful crawl and a sys.exit(1) on a failed crawl to indicate when a container is finished to the Docker managing script, which listens to the Docker event bus via container.wait()
Removed redundant multi_threaded watcher. Crawler will simply fail for a given page if it can not load it within 3 tries, and will move on to the next.